fix init issue?
