+++ src/tests/attestation.test.ts
@@ -1,10 +1,12 @@
 import "./setup"; // Must be imported first to set up environment variables
 import { describe, expect, it } from "bun:test";
-import app from "../index";
+import { createApp } from "../app";
 import { env } from "../config";
 
 describe("Attestation", () => {
 	it("should return 200 and a public key", async () => {
+		const app = await createApp();
+
 		const req = new Request("http://localhost/v1/attestation/public-key", {
 			headers: {
 				authorization: env.ACCESS_SECRET,
