+++ src/tests/healthcheck.test.ts
@@ -1,10 +1,12 @@
 import "./setup"; // Must be imported first to set up environment variables
 import { describe, expect, it } from "bun:test";
-import app from "../index";
+import { createApp } from "../app";
 import { env } from "../config";
 
 describe("Healthcheck", () => {
 	it("should return 200", async () => {
+		const app = await createApp();
+
 		const req = new Request("http://localhost/health", {
 			headers: {
 				authorization: env.ACCESS_SECRET,
